Added fixes

+++ Domain/Entities/Play.cs
@@ -7,12 +7,12 @@ public class Play
     public string Name { get; private set; }
     public DateTime StartDateTimeUtc { get; private set; }
     public DateTime EndDateTimeUtc { get; private set; }
-    public int Price { get; private set; }
+    public decimal Price { get; private set; }
     public string Description { get; private set; }
     public int TheaterId { get; private init; }
     public int CompositionId { get; private init; }
 
-    public Play( int playId, string name, DateTime startDateTimeUtc, DateTime endDateTimeUtc, int price, string description, int theaterId, int compositionId )
+    public Play( int playId, string name, DateTime startDateTimeUtc, DateTime endDateTimeUtc, decimal price, string description, int theaterId, int compositionId )
     {
         PlayId = playId;
         Name = name;
